Fix interactive widgets sometimes being incorrectly marked as hovered
An interactive widget should only be marked hovered if a click/drag
would start an interaction with it.

egui 0.30 introduced a feature where a thin interactive widget
could be hit even if it was partially behind a larger interactive
widget. Unfortunately, this introduced a bug where the top widget
would still be marked as hovered, even though a click would go through
to the thin widget below.

This bug was most notacible when trying to reisize a window by dragging
its corner, which often would result in dragging one of its sides
instead.

This PR fixes this bug.