Stop wrapping FontsImpl in an Arc<Mutex<_>>
Any attempt to actually borrow it multiple times would deadlock, we
never need to clone it, and none of its methods took `self` as mutable,
meaning it wasn't making use of the semantic difference between the two.

This API is about to get reworked, and simplifying it is a first step.