Rework modal drawing to better use provided Area
When creating a modal with a custom Area that is `Area::is_movable`, i
would expect to be able to drag it around.

However, that is not the case.

Modal, in rough terms, currently does the following:
 * Enter the area using `Area::show()`
 * Draw the backdrop and get its reponse
 * Create a child ui to stop events from going to the backdrop, as that
   might cause clicks inside the frame below to close the modal.
 * Draw the frame, and in it the user content.

The main issue this creates is that drawing the backdrop after entering
the area via `Area::show()` captures the events before they hit the
area, and thus a moveable area cannot be dragged - it does not see any
events to itself in the first place.

Instead, i have modified `Modal::show` to do the following
 * Create the backdrop response using `Context::create_widget`.
 * Then enter the area.
 * Draw the frame, and then the user content.

Note that the child ui is now not required anymore, since as long as the
Area senses click and drag, it eats those events now.

I was then left with two issues:
  * Modal::default_area() cannot ever become a draggable thing (even if
    setting movable to true), as it is anchored, and i cannot reset the
    anchor.
  * If i create a custom Area to be moved around, it is not centered
    anymore, as without an anchor it will just get a auto-position.

The solution for the first problem is that i have
  * Added documentation to `Modal::area` and `Modal::default_area`
  * Added a new function, `Modal::draggable_area`, to give a quick area
    that can be dragged around and behaves as expected.

The second problem is a bit trickier. Will a user want an arbirtrary
position for the modal? If yes, i expect they would set fixed_position,
right? If thats the case, there is no harm in re-centering the modal
every time it is first shown. I query that using the areas id and
`Memory::areas().visible_last_frame()` and it appears to be working
reasonably well.

If its however expected that the user will use `default_pos`,
currently that will "break" (as in, defaukt_pos gets ignored) as i have not find a way to query it in `Modals` code if
`default_pos` was set. Of course that is just an added function to
`Area` away, but i didn't want to expand the scope of this.

But for now i think this solution is a sufficient compromise.

Finally, i have also added code to the egui demo lib to demonstrate a draggable
and resizable modal.